fix: change test to jest
