+++ package.json
@@ -21,6 +21,7 @@
     "findup": "^0.1.5",
     "ghooks": "^1.3.2",
     "glob": "^7.1.1",
+    "image-size": "^0.5.1",
     "istanbul-api": "^1.0.0-aplha.10",
     "istanbul-lib-coverage": "^1.0.0",
     "jest": "^18.1.0",
