+++ packages/image-source-loader/package.json
@@ -2,12 +2,13 @@
   "name": "image-source-loader",
   "version": "0.1.1",
   "description": "image source loader for webpack",
+  "license": "BSD-3-Clause",
   "repository": {
     "type": "git",
     "url": "https://github.com/alibaba/rax.git"
   },
-  "scripts": {
-    "test": "mocha"
+  "engines": {
+    "npm": ">=3.0.0"
   },
   "keywords": [
     "base64",
@@ -18,16 +19,9 @@
     "src",
     "img src"
   ],
-  "main": "index.js",
+  "main": "lib/index.js",
   "author": "wssgcg1213 <i@zeroling.com>",
   "license": "MIT",
-  "devDependencies": {
-    "chai": "^3.5.0",
-    "eslint": "^2.11.0",
-    "eslint-config-airbnb-base": "^3.0.1",
-    "eslint-plugin-import": "^1.8.1",
-    "mocha": "^2.5.3"
-  },
   "dependencies": {
     "image-size": "^0.5.1"
   }
