--- packages/image-source-loader/test/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "parserOptions": {
-    "sourceType": "module"
-  }
-}
